requirements.txt
aioredis==1.3.1
aiosmtplib==1.1.6
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
charset-normalizer==2.1.1
config==0.5.0.post0
constantly==15.1.0
cryptography==37.0.1
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.15
django-admin-bootstrapped==2.5.7
django-allauth==0.51.0
django-appconf==1.0.5
django-autoscroll==1.0.0
django-bootstrap==0.2.4
django-bootstrap4==22.2
django-cleanup==6.0.0
django-compressor==4.0
django-cors-headers==3.2.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-libsass==0.9
django-sass-processor==1.1
django-templated-mail==1.1.1
djangochannelsrestframework==0.3.0
djangorestframework==3.12.4
djangorestframework-jsonapi==3.0.0
djoser==2.0.3
docker==6.0.1
docopt==0.6.2
et-xmlfile==1.1.0
h11==0.12.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
hypercorn==0.12.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==2.1.3
incremental==21.3.0
inflection==0.3.1
libsass==0.21.0
msgpack==1.0.3
mysql-client==0.0.1
oauthlib==3.1.0
openpyxl==3.0.10
packaging==21.3
Pillow==9.1.1
pipreqs==0.4.11
priority==2.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pygame==2.0.1
PyHamcrest==1.9.0
PyJWT==2.3.0
Pyntifex==0.0.14
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-Qt5==5.15.2
PySide2==5.15.2.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
pywin32==305
rcssmin==1.1.0
redis==4.3.5
requests==2.22.0
requests-oauthlib==1.3.0
rjsmin==1.2.0
service-identity==21.1.0
shiboken2==5.15.2.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.8.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2021.5
urllib3==1.25.7
websocket-client==1.4.2
wsproto==1.0.0
yarg==0.1.9
zipp==3.10.0
zope.interface==5.4.0
